:recycle: changing folder structures to use the DDD Design Pattern
