+++ backend/src/modules/users/infra/http/middlewares/ensuredAuthenticated.ts
@@ -1,9 +1,9 @@
 import { Request, Response, NextFunction } from 'express';
 import { verify } from 'jsonwebtoken';
 
-import AppError from '../errors/AppError';
+import AppError from '@shared/errors/AppError';
 
-import authConfig from '../config/auth';
+import authConfig from '@config/auth';
 
 interface TokenPayload {
   iat: number;
