+++ backend/src/modules/users/infra/http/routes/departament.routes.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express';
 import { getRepository } from 'typeorm';
 
-import AppError from '../errors/AppError';
-import Departament from '../entities/Departament';
-import idValidedUiid from '../middlewares/idValidedUuid';
-import CreateDepartamentService from '../services/CreateDepartamentService';
-import UpdateDepartamentService from '../services/UpdateDepartamentService';
+import AppError from '@shared/errors/AppError';
+import idValidedUiid from '@shared/infra/http/middlewares/idValidedUuid';
+import Departament from '@modules/users/infra/typeorm/entities/Departament';
+import CreateDepartamentService from '@modules/users/services/CreateDepartamentService';
+import UpdateDepartamentService from '@modules/users/services/UpdateDepartamentService';
 
 const departamentRouter = Router();
 
@@ -15,7 +15,6 @@ departamentRouter.get('/', async (request, response) => {
 
   return response.json(departaments);
 });
-
 
 departamentRouter.post('/', async (request, response) => {
   const { name, initials, goal, subsidiary_id } = request.body;
@@ -31,7 +30,6 @@ departamentRouter.post('/', async (request, response) => {
   return response.json(departament);
 });
 
-
 departamentRouter.get('/:id', idValidedUiid, async (request, response) => {
   const departamentRepository = getRepository(Departament);
   const departament = await departamentRepository.findOne(request.params.id);
@@ -43,7 +41,6 @@ departamentRouter.get('/:id', idValidedUiid, async (request, response) => {
   return response.json(departament);
 });
 
-
 departamentRouter.put('/:id', idValidedUiid, async (request, response) => {
   const updateDepartamentService = new UpdateDepartamentService();
 
@@ -54,7 +51,6 @@ departamentRouter.put('/:id', idValidedUiid, async (request, response) => {
 
   return response.json(departamentUpdated);
 });
-
 
 departamentRouter.delete('/:id', idValidedUiid, async (request, response) => {
   const departamentRepository = getRepository(Departament);
