+++ backend/src/modules/users/infra/http/routes/session.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 
-import SessionController from '../controllers/SessionController';
+import SessionController from '@modules/users/infra/http/controllers/SessionController';
 
 const sessionsRouter = Router();
 const sessionController = new SessionController();
