+++ backend/src/modules/users/infra/http/routes/user.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 
-import UsersController from '../controllers/UsersController';
-import idValidUuid from '../middlewares/idValidedUuid';
+import UsersController from '@modules/users/infra/http/controllers/UsersController';
+import idValidUuid from '@shared/infra/http/middlewares/idValidedUuid';
 
 const usersRouter = Router();
 const usersController = new UsersController();
