+++ backend/src/shared/infra/http/server.ts
@@ -2,10 +2,10 @@ import 'reflect-metadata';
 import express, { Request, Response, NextFunction } from 'express';
 import 'express-async-errors';
 
-import routes from './routes';
-import AppError from './errors/AppError';
+import routes from '@shared/infra/http/routes';
+import AppError from '@shared/errors/AppError';
 
-import './database';
+import '@shared/infra/typeorm';
 
 const app = express();
 app.use(express.json());
